+++ src/kihachi_mcp/models/__init__.py
@@ -1,3 +1,8 @@
+from kihachi_mcp.models.ableton_plan import (
+    AbletonLocator,
+    AbletonProjectPlan,
+    AbletonTrackPlan,
+)
 from kihachi_mcp.models.arrangement import Arrangement
 from kihachi_mcp.models.genre_template import GenreTemplate
 from kihachi_mcp.models.project_plan import ProjectPlan
@@ -6,6 +11,9 @@ from kihachi_mcp.models.songspec import SongSpec
 from kihachi_mcp.models.track import TrackSpec
 
 __all__ = [
+    "AbletonLocator",
+    "AbletonProjectPlan",
+    "AbletonTrackPlan",
     "Arrangement",
     "GenreTemplate",
     "ProjectPlan",