+++ src/kihachi_mcp/tools/__init__.py
@@ -1,8 +1,10 @@
+from kihachi_mcp.tools.ableton import create_ableton_plan
 from kihachi_mcp.tools.brain import generate_songspec
 from kihachi_mcp.tools.hello import hello
 from kihachi_mcp.tools.project_builder import create_project_from_songspec
 
 __all__ = [
+    "create_ableton_plan",
     "create_project_from_songspec",
     "generate_songspec",
     "hello",